refactor: migrate AppConfig to pydantic-settings and relocate from services
- Replace custom AppConfig implementation with pydantic-settings BaseSettings
- Move config from services/config_service.py to app_config.py at package root
- Add pydantic-settings dependency
- Use NoDecode annotation for comma-separated list fields
- Add field validator for parsing comma-separated integers
- Update all imports throughout the codebase
- Simplify config instantiation with factory function
- Update compose files to use consistent environment variables
