+++ compose-local.yaml
@@ -11,6 +11,8 @@ services:
     environment:
       OPENAI_MODEL: ${OPENAI_MODEL}
       OPENAI_API_KEY: ${OPENAI_API_KEY}
+      ANTHROPIC_API_KEY: ${ANTHROPIC_API_KEY}
+      ANTHROPIC_MODEL: ${ANTHROPIC_MODEL}
       TELEGRAM_BOT_TOKEN: ${TELEGRAM_BOT_TOKEN}
       TELEGRAM_API_ID: ${TELEGRAM_API_ID}
       TELEGRAM_API_HASH: ${TELEGRAM_API_HASH}
@@ -18,8 +20,10 @@ services:
       TELEGRAM_ALLOWED_USERS: ${TELEGRAM_ALLOWED_USERS}
       USER_IDS_TO_SEND_MORNING_MSG: ${USER_IDS_TO_SEND_MORNING_MSG}
       CHAT_IDS_TO_SEND_STORIES: ${CHAT_IDS_TO_SEND_STORIES}
-      REDIS_URL: ${REDIS_URL}
-      PHOENIX_COLLECTOR_ENDPOINT: ${PHOENIX_COLLECTOR_ENDPOINT}
+      PHOENIX_COLLECTOR_ENDPOINT: http://phoenix:6006
+      REDIS_URL: redis://story-cache:6379/0
+    networks:
+      - telemetry
     volumes:
       - ./little_turtle/services/little_turtle.session:/app/little_turtle/services/little_turtle.session
       - story-images:/app/little_turtle/images
