+++ compose.yaml
@@ -6,6 +6,8 @@ services:
     environment:
       OPENAI_MODEL: ${OPENAI_MODEL}
       OPENAI_API_KEY: ${OPENAI_API_KEY}
+      ANTHROPIC_API_KEY: ${ANTHROPIC_API_KEY}
+      ANTHROPIC_MODEL: ${ANTHROPIC_MODEL}
       TELEGRAM_BOT_TOKEN: ${TELEGRAM_BOT_TOKEN}
       TELEGRAM_API_ID: ${TELEGRAM_API_ID}
       TELEGRAM_API_HASH: ${TELEGRAM_API_HASH}
@@ -13,8 +15,8 @@ services:
       TELEGRAM_ALLOWED_USERS: ${TELEGRAM_ALLOWED_USERS}
       USER_IDS_TO_SEND_MORNING_MSG: ${USER_IDS_TO_SEND_MORNING_MSG}
       CHAT_IDS_TO_SEND_STORIES: ${CHAT_IDS_TO_SEND_STORIES}
-      REDIS_URL: ${REDIS_URL}
-      PHOENIX_COLLECTOR_ENDPOINT: ${PHOENIX_COLLECTOR_ENDPOINT}
+      REDIS_URL: redis://story-cache:6379/0
+      PHOENIX_COLLECTOR_ENDPOINT: http://phoenix:6006
     volumes:
       - ./little_turtle/services/little_turtle.session:/app/little_turtle/services/little_turtle.session
       - story-images:/app/little_turtle/images
