+++ little_turtle/services/__init__.py
@@ -1,9 +1,7 @@
-from .config_service import AppConfig
 from .logger_service import LoggerService
 from .telegram_service import TelegramService
 
 __all__ = [
-    "AppConfig",
     "LoggerService",
     "TelegramService",
 ]
