+++ pyproject.toml
@@ -21,6 +21,7 @@ dependencies = [
     "openinference-instrumentation-openai>=0.1.30",
     "openinference-instrumentation-anthropic>=0.1.18",
     "arize-phoenix-client>=1.13.2",
+    "pydantic-settings>=2.10.1",
 ]
 
 [dependency-groups]
